Changed from using numpy dot and numpy inverse methods in train_LM function to scipy alternatives.
The first change I found useful was switching from using the
numpy.dot method to the scipy alternative called
scipy.linalg.blas.dgemm method. The reason for this change is
under the hood when you call numpy.dot, it copies the array's due
to its need to have them in C-contiguous order, but when you
switch to the scipy alternative, it doesn't copy the arrays as you
pass them in as C-contiguous. This change both prevents the excess
memory consumption of the copy, and also the performance bottleneck
due to it. This change helps mostly when it comes to very large
datasets such as the ones i'm using and the result is significantly
faster using the scipy version.

The second possible change I found useful was switching from the
numpy inverse method to its scipy alternative. The reason for this
is scipy always is compiled with BLAS/LAPACK support while for
numpy it is only optional. I found utilizing this change resulted
in this calculation taking 1/6th the amount of time roughly when
using very large datasets.

Feel free to test these changes and let me know if you like them.
Also thank you for providing this code as it has helped me on a
personal project of mine.